requirements.txt
aiofiles==23.2.1
amqp==5.3.1
annotated-types==0.6.0
anthropic==0.60.0
anyio==3.7.1
asttokens==3.0.0
bcrypt==4.3.0
billiard==4.2.1
black==23.11.0
celery==5.3.4
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==45.0.5
decorator==5.2.1
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
executing==2.2.0
fastapi==0.104.1
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.25.2
idna==3.10
iniconfig==2.1.0
ipython==9.4.0
ipython_pygments_lexers==1.1.1
isort==5.12.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
kombu==5.5.4
langchain-anthropic==0.3.18
langchain-core==0.3.72
langgraph==0.5.4
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.5.2
langgraph-sdk==0.1.74
langsmith==0.4.8
loguru==0.7.2
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
motor==3.3.2
mypy==1.7.1
mypy_extensions==1.1.0
orjson==3.11.0
ormsgpack==1.10.0
packaging==25.0
parso==0.8.4
passlib==1.7.4
pathspec==0.12.1
pexpect==4.9.0
pip==25.1
platformdirs==4.3.8
pluggy==1.6.0
prometheus-client==0.19.0
prompt_toolkit==3.0.51
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pycparser==2.22
pydantic==2.7.4
pydantic_core==2.18.4
pydantic-settings==2.0.3
Pygments==2.19.2
pymongo==4.6.0
pytest==7.4.3
pytest-asyncio==0.21.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.2
redis==5.0.1
requests==2.32.4
requests-toolbelt==1.0.0
rsa==4.9.1
sentry-sdk==1.38.0
setuptools==78.1.1
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.27.0
structlog==23.2.0
tenacity==9.1.2
traitlets==5.14.3
typing_extensions==4.14.1
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.24.0
uvloop==0.21.0
vine==5.1.0
watchfiles==1.1.0
wcwidth==0.2.13
websockets==15.0.1
wheel==0.45.1
xxhash==3.5.0
zstandard==0.23.0